move call forest code next to account update
