+++ src/lib/account_update.ts
@@ -3,6 +3,8 @@ import {
   FlexibleProvable,
   provable,
   provablePure,
+  Struct,
+  Unconstrained,
 } from './circuit_value.js';
 import { memoizeWitness, Provable } from './provable.js';
 import { Field, Bool } from './core.js';
@@ -33,18 +35,29 @@ import {
   Actions,
 } from '../bindings/mina-transaction/transaction-leaves.js';
 import { TokenId as Base58TokenId } from './base58-encodings.js';
-import { hashWithPrefix, packToFields } from './hash.js';
+import {
+  hashWithPrefix,
+  packToFields,
+  Poseidon,
+  ProvableHashable,
+} from './hash.js';
 import { mocks, prefixes } from '../bindings/crypto/constants.js';
 import { Context } from './global-context.js';
 import { Signature, signFieldElement } from '../mina-signer/src/signature.js';
 import { transactionCommitments } from '../mina-signer/src/sign-zkapp-command.js';
-import type { CallForestUnderConstruction } from './mina/token/call-forest.js';
 import { currentTransaction } from './mina/transaction-context.js';
 import { isSmartContract } from './mina/smart-contract-base.js';
 import { activeInstance } from './mina/mina-instance.js';
+import {
+  genericHash,
+  MerkleList,
+  MerkleListBase,
+  withHashes,
+} from './provable-types/merkle-list.js';
+import { Hashed } from './provable-types/packed.js';
 
 // external API
-export { AccountUpdate, Permissions, ZkappPublicInput };
+export { AccountUpdate, Permissions, ZkappPublicInput, CallForest };
 // internal API
 export {
   smartContractContext,
@@ -61,7 +74,7 @@ export {
   Actions,
   TokenId,
   Token,
-  CallForest,
+  CallForestHelpers,
   createChildAccountUpdate,
   AccountUpdatesLayout,
   zkAppProver,
@@ -70,6 +83,9 @@ export {
   dummySignature,
   LazySignature,
   LazyProof,
+  CallTree,
+  CallForestUnderConstruction,
+  hashAccountUpdate,
 };
 
 const ZkappStateLength = 8;
@@ -961,7 +977,7 @@ class AccountUpdate implements Types.AccountUpdate {
     if (isSameAsFeePayer) nonce++;
     // now, we check how often this account update already updated its nonce in
     // this tx, and increase nonce from `getAccount` by that amount
-    CallForest.forEachPredecessor(
+    CallForestHelpers.forEachPredecessor(
       currentTransaction.get().accountUpdates,
       update as AccountUpdate,
       (otherUpdate) => {
@@ -1008,7 +1024,7 @@ class AccountUpdate implements Types.AccountUpdate {
 
   toPublicInput(): ZkappPublicInput {
     let accountUpdate = this.hash();
-    let calls = CallForest.hashChildren(this);
+    let calls = CallForestHelpers.hashChildren(this);
     return { accountUpdate, calls };
   }
 
@@ -1284,7 +1300,7 @@ class AccountUpdate implements Types.AccountUpdate {
     if (n === 0) {
       accountUpdate.children.callsType = {
         type: 'Equals',
-        value: CallForest.emptyHash(),
+        value: CallForestHelpers.emptyHash(),
       };
     }
   }
@@ -1483,7 +1499,148 @@ type AccountUpdatesLayout =
   | 'NoDelegation'
   | AccountUpdatesLayout[];
 
-const CallForest = {
+// call forest stuff
+
+function hashAccountUpdate(update: AccountUpdate) {
+  return genericHash(AccountUpdate, prefixes.body, update);
+}
+
+class HashedAccountUpdate extends Hashed.create(
+  AccountUpdate,
+  hashAccountUpdate
+) {}
+
+type CallTree = {
+  accountUpdate: Hashed<AccountUpdate>;
+  calls: MerkleListBase<CallTree>;
+};
+const CallTree: ProvableHashable<CallTree> = Struct({
+  accountUpdate: HashedAccountUpdate.provable,
+  calls: MerkleListBase<CallTree>(),
+});
+
+class CallForest extends MerkleList.create(CallTree, merkleListHash) {
+  static fromAccountUpdates(updates: AccountUpdate[]): CallForest {
+    let nodes = updates.map((update) => {
+      let accountUpdate = HashedAccountUpdate.hash(update);
+      let calls = CallForest.fromAccountUpdates(update.children.accountUpdates);
+      return { accountUpdate, calls };
+    });
+
+    return CallForest.from(nodes);
+  }
+}
+
+// how to hash a forest
+
+function merkleListHash(forestHash: Field, tree: CallTree) {
+  return hashCons(forestHash, hashNode(tree));
+}
+function hashNode(tree: CallTree) {
+  return Poseidon.hashWithPrefix(prefixes.accountUpdateNode, [
+    tree.accountUpdate.hash,
+    tree.calls.hash,
+  ]);
+}
+function hashCons(forestHash: Field, nodeHash: Field) {
+  return Poseidon.hashWithPrefix(prefixes.accountUpdateCons, [
+    nodeHash,
+    forestHash,
+  ]);
+}
+
+/**
+ * Structure for constructing a call forest from a circuit.
+ *
+ * The circuit can mutate account updates and change their array of children, so here we can't hash
+ * everything immediately. Instead, we maintain a structure consisting of either hashes or full account
+ * updates that can be hashed into a final call forest at the end.
+ */
+type CallForestUnderConstruction = HashOrValue<
+  {
+    accountUpdate: HashOrValue<AccountUpdate>;
+    calls: CallForestUnderConstruction;
+  }[]
+>;
+
+type HashOrValue<T> =
+  | { useHash: true; hash: Field; value: T }
+  | { useHash: false; value: T };
+
+const CallForestUnderConstruction = {
+  empty(): CallForestUnderConstruction {
+    return { useHash: false, value: [] };
+  },
+
+  setHash(forest: CallForestUnderConstruction, hash: Field) {
+    Object.assign(forest, { useHash: true, hash });
+  },
+
+  witnessHash(forest: CallForestUnderConstruction) {
+    let hash = Provable.witness(Field, () => {
+      let nodes = forest.value.map(toCallTree);
+      return withHashes(nodes, merkleListHash).hash;
+    });
+    CallForestUnderConstruction.setHash(forest, hash);
+  },
+
+  push(
+    forest: CallForestUnderConstruction,
+    accountUpdate: HashOrValue<AccountUpdate>,
+    calls?: CallForestUnderConstruction
+  ) {
+    forest.value.push({
+      accountUpdate,
+      calls: calls ?? CallForestUnderConstruction.empty(),
+    });
+  },
+
+  remove(forest: CallForestUnderConstruction, accountUpdate: AccountUpdate) {
+    // find account update by .id
+    let index = forest.value.findIndex(
+      (node) => node.accountUpdate.value.id === accountUpdate.id
+    );
+
+    // nothing to do if it's not there
+    if (index === -1) return;
+
+    // remove it
+    forest.value.splice(index, 1);
+  },
+
+  finalize(forest: CallForestUnderConstruction): CallForest {
+    if (forest.useHash) {
+      let data = Unconstrained.witness(() => {
+        let nodes = forest.value.map(toCallTree);
+        return withHashes(nodes, merkleListHash).data;
+      });
+      return new CallForest({ hash: forest.hash, data });
+    }
+
+    // not using the hash means we calculate it in-circuit
+    let nodes = forest.value.map(toCallTree);
+    return CallForest.from(nodes);
+  },
+};
+
+function toCallTree(node: {
+  accountUpdate: HashOrValue<AccountUpdate>;
+  calls: CallForestUnderConstruction;
+}): CallTree {
+  let accountUpdate = node.accountUpdate.useHash
+    ? new HashedAccountUpdate(
+        node.accountUpdate.hash,
+        Unconstrained.from(node.accountUpdate.value)
+      )
+    : HashedAccountUpdate.hash(node.accountUpdate.value);
+
+  return {
+    accountUpdate,
+    calls: CallForestUnderConstruction.finalize(node.calls),
+  };
+}
+
+const CallForestHelpers = {
   // similar to Mina_base.ZkappCommand.Call_forest.to_account_updates_list
   // takes a list of accountUpdates, which each can have children, so they form a "forest" (list of trees)
   // returns a flattened list, with `accountUpdate.body.callDepth` specifying positions in the forest
@@ -1500,7 +1657,7 @@ const CallForest = {
       let children = accountUpdate.children.accountUpdates;
       accountUpdates.push(
         accountUpdate,
-        ...CallForest.toFlatList(children, mutate, depth + 1)
+        ...CallForestHelpers.toFlatList(children, mutate, depth + 1)
       );
     }
     return accountUpdates;
@@ -1516,19 +1673,21 @@ const CallForest = {
   // the `calls` field of ZkappPublicInput
   hashChildren(update: AccountUpdate): Field {
     if (!Provable.inCheckedComputation()) {
-      return CallForest.hashChildrenBase(update);
+      return CallForestHelpers.hashChildrenBase(update);
     }
 
     let { callsType } = update.children;
     // compute hash outside the circuit if callsType is "Witness"
     // i.e., allowing accountUpdates with arbitrary children
     if (callsType.type === 'Witness') {
-      return Provable.witness(Field, () => CallForest.hashChildrenBase(update));
+      return Provable.witness(Field, () =>
+        CallForestHelpers.hashChildrenBase(update)
+      );
     }
     if (callsType.type === 'WitnessEquals') {
       return callsType.value;
     }
-    let calls = CallForest.hashChildrenBase(update);
+    let calls = CallForestHelpers.hashChildrenBase(update);
     if (callsType.type === 'Equals') {
       calls.assertEquals(callsType.value);
     }
@@ -1536,9 +1695,9 @@ const CallForest = {
   },
 
   hashChildrenBase({ children }: AccountUpdate) {
-    let stackHash = CallForest.emptyHash();
+    let stackHash = CallForestHelpers.emptyHash();
     for (let accountUpdate of [...children.accountUpdates].reverse()) {
-      let calls = CallForest.hashChildren(accountUpdate);
+      let calls = CallForestHelpers.hashChildren(accountUpdate);
       let nodeHash = hashWithPrefix(prefixes.accountUpdateNode, [
         accountUpdate.hash(),
         calls,
@@ -1564,7 +1723,7 @@ const CallForest = {
     let newUpdates: AccountUpdate[] = [];
     for (let update of updates) {
       let newUpdate = map(update);
-      newUpdate.children.accountUpdates = CallForest.map(
+      newUpdate.children.accountUpdates = CallForestHelpers.map(
         update.children.accountUpdates,
         map
       );
@@ -1576,7 +1735,7 @@ const CallForest = {
   forEach(updates: AccountUpdate[], callback: (update: AccountUpdate) => void) {
     for (let update of updates) {
       callback(update);
-      CallForest.forEach(update.children.accountUpdates, callback);
+      CallForestHelpers.forEach(update.children.accountUpdates, callback);
     }
   },
 
@@ -1586,7 +1745,7 @@ const CallForest = {
     callback: (update: AccountUpdate) => void
   ) {
     let isPredecessor = true;
-    CallForest.forEach(updates, (otherUpdate) => {
+    CallForestHelpers.forEach(updates, (otherUpdate) => {
       if (otherUpdate.id === update.id) isPredecessor = false;
       if (isPredecessor) callback(otherUpdate);
     });
