+++ src/lib/mina/token/call-forest.ts
@@ -1,48 +1,11 @@
-import { prefixes } from '../../../provable/poseidon-bigint.js';
-import { AccountUpdate, TokenId } from '../../account_update.js';
+import { AccountUpdate, CallForest, TokenId } from '../../account_update.js';
 import { Field } from '../../core.js';
 import { Provable } from '../../provable.js';
-import { Struct, Unconstrained } from '../../circuit_value.js';
+import { Struct } from '../../circuit_value.js';
 import { assert } from '../../gadgets/common.js';
-import { Poseidon, ProvableHashable } from '../../hash.js';
-import { Hashed } from '../../provable-types/packed.js';
-import {
-  MerkleArray,
-  MerkleListBase,
-  MerkleList,
-  genericHash,
-  withHashes,
-} from '../../provable-types/merkle-list.js';
+import { MerkleArray, MerkleList } from '../../provable-types/merkle-list.js';
 
-export { CallForest, CallForestArray, CallForestIterator, hashAccountUpdate };
-
-export { CallForestUnderConstruction };
-
-class HashedAccountUpdate extends Hashed.create(
-  AccountUpdate,
-  hashAccountUpdate
-) {}
-
-type CallTree = {
-  accountUpdate: Hashed<AccountUpdate>;
-  calls: MerkleListBase<CallTree>;
-};
-const CallTree: ProvableHashable<CallTree> = Struct({
-  accountUpdate: HashedAccountUpdate.provable,
-  calls: MerkleListBase<CallTree>(),
-});
-
-class CallForest extends MerkleList.create(CallTree, merkleListHash) {
-  static fromAccountUpdates(updates: AccountUpdate[]): CallForest {
-    let nodes = updates.map((update) => {
-      let accountUpdate = HashedAccountUpdate.hash(update);
-      let calls = CallForest.fromAccountUpdates(update.children.accountUpdates);
-      return { accountUpdate, calls };
-    });
-
-    return CallForest.from(nodes);
-  }
-}
+export { CallForestArray, CallForestIterator };
 
 class CallForestArray extends MerkleArray.createFromList(CallForest) {}
 
@@ -157,118 +120,4 @@ class CallForestIterator {
       message ?? 'CallForest not finished'
     );
   }
-}
-
-// how to hash a forest
-
-function merkleListHash(forestHash: Field, tree: CallTree) {
-  return hashCons(forestHash, hashNode(tree));
-}
-
-function hashNode(tree: CallTree) {
-  return Poseidon.hashWithPrefix(prefixes.accountUpdateNode, [
-    tree.accountUpdate.hash,
-    tree.calls.hash,
-  ]);
-}
-function hashCons(forestHash: Field, nodeHash: Field) {
-  return Poseidon.hashWithPrefix(prefixes.accountUpdateCons, [
-    nodeHash,
-    forestHash,
-  ]);
-}
-
-function hashAccountUpdate(update: AccountUpdate) {
-  return genericHash(AccountUpdate, prefixes.body, update);
-}
-
-/**
- * Structure for constructing a call forest from a circuit.
- *
- * The circuit can mutate account updates and change their array of children, so here we can't hash
- * everything immediately. Instead, we maintain a structure consisting of either hashes or full account
- * updates that can be hashed into a final call forest at the end.
- */
-type CallForestUnderConstruction = HashOrValue<
-  {
-    accountUpdate: HashOrValue<AccountUpdate>;
-    calls: CallForestUnderConstruction;
-  }[]
->;
-
-type HashOrValue<T> =
-  | { useHash: true; hash: Field; value: T }
-  | { useHash: false; value: T };
-
-const CallForestUnderConstruction = {
-  empty(): CallForestUnderConstruction {
-    return { useHash: false, value: [] };
-  },
-
-  setHash(forest: CallForestUnderConstruction, hash: Field) {
-    Object.assign(forest, { useHash: true, hash });
-  },
-
-  witnessHash(forest: CallForestUnderConstruction) {
-    let hash = Provable.witness(Field, () => {
-      let nodes = forest.value.map(toCallTree);
-      return withHashes(nodes, merkleListHash).hash;
-    });
-    CallForestUnderConstruction.setHash(forest, hash);
-  },
-
-  push(
-    forest: CallForestUnderConstruction,
-    accountUpdate: HashOrValue<AccountUpdate>,
-    calls?: CallForestUnderConstruction
-  ) {
-    forest.value.push({
-      accountUpdate,
-      calls: calls ?? CallForestUnderConstruction.empty(),
-    });
-  },
-
-  remove(forest: CallForestUnderConstruction, accountUpdate: AccountUpdate) {
-    // find account update by .id
-    let index = forest.value.findIndex(
-      (node) => node.accountUpdate.value.id === accountUpdate.id
-    );
-
-    // nothing to do if it's not there
-    if (index === -1) return;
-
-    // remove it
-    forest.value.splice(index, 1);
-  },
-
-  finalize(forest: CallForestUnderConstruction): CallForest {
-    if (forest.useHash) {
-      let data = Unconstrained.witness(() => {
-        let nodes = forest.value.map(toCallTree);
-        return withHashes(nodes, merkleListHash).data;
-      });
-      return new CallForest({ hash: forest.hash, data });
-    }
-
-    // not using the hash means we calculate it in-circuit
-    let nodes = forest.value.map(toCallTree);
-    return CallForest.from(nodes);
-  },
-};
-
-function toCallTree(node: {
-  accountUpdate: HashOrValue<AccountUpdate>;
-  calls: CallForestUnderConstruction;
-}): CallTree {
-  let accountUpdate = node.accountUpdate.useHash
-    ? new HashedAccountUpdate(
-        node.accountUpdate.hash,
-        Unconstrained.from(node.accountUpdate.value)
-      )
-    : HashedAccountUpdate.hash(node.accountUpdate.value);
-
-  return {
-    accountUpdate,
-    calls: CallForestUnderConstruction.finalize(node.calls),
-  };
 }
